Move LNURL-withdraw request, error into sdk-common
