+++ libs/sdk-core/src/breez_services.rs
@@ -32,7 +32,7 @@ use crate::chain::{
     DEFAULT_MEMPOOL_SPACE_URL,
 };
 use crate::error::{
-    ConnectError, LnUrlAuthError, LnUrlWithdrawError, ReceiveOnchainError, ReceiveOnchainResult,
+    ConnectError, LnUrlAuthError, ReceiveOnchainError, ReceiveOnchainResult,
     ReceivePaymentError, SdkError, SdkResult, SendOnchainError, SendPaymentError,
 };
 use crate::fiat::{FiatCurrency, Rate};
