+++ libs/sdk-core/src/bridge_generated.rs
@@ -53,7 +53,6 @@ use crate::models::GreenlightNodeConfig;
 use crate::models::HealthCheckStatus;
 use crate::models::ListPaymentsRequest;
 use crate::models::LnPaymentDetails;
-use crate::models::LnUrlWithdrawRequest;
 use crate::models::LogEntry;
 use crate::models::MaxReverseSwapAmountResponse;
 use crate::models::MetadataFilter;
@@ -115,6 +114,7 @@ use crate::LnUrlErrorData;
 use crate::LnUrlPayErrorData;
 use crate::LnUrlPayRequest;
 use crate::LnUrlPayRequestData;
+use crate::LnUrlWithdrawRequest;
 use crate::LnUrlWithdrawRequestData;
 use crate::LnUrlWithdrawResult;
 use crate::LnUrlWithdrawSuccessData;
